# routes/swagger.yaml
...
swagger: '2.0'
info:
  title: KipsWarehouse API
  description: Move your app forward with the KipsWarehouse API
  version: "1.0.0"
# the domain of the service
host: kipswarehouse.com
...
schemes:
  - https
securityDefinitions:
  JWT:
    type: apiKey
    name: Authorization
    in: header
basePath: /
produces:
  - application/json
paths:
...
  /api/items/:
    get:
      security:
        - JWT: []
      description: |
        Get all the items with proper filtering
      parameters:
        - name: includeTags
          in: query
          description: json array of tags to be included in the search
          required: false
          type: string
        - name: excludeTags
          in: query
          description: json array of tags to be excluded in the search
          required: false
          type: string
        - name: id
          in: query
          description: identifies an item
          required: false
          type: number
          format: integer
        - name: rowPerPage
          in: query
          description: the number of rows in a page for pagination
          required: false
          type: number
          format: integer
        - name: pageNumber
          in: query
          description: starting counting at 1
          required: false
          type: number
          format: integer
        - name: name
          in: query
          description: the fizzy name of an item to be searched
          required: false
          type: string
        - name: model
          in: query
          description: the fizzy model of an item to be searched
          required: false
          type: string
      tags:
        - Items
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /api/items/customFields:
    get:
      security:
        - JWT: []
      description: get all custom fields defined on items
      tags:
        - Custom Fields
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /api/items/customFields/{name}/{type}/{visibility} :
    post:
      security:
        - JWT: []
      description: Create this custom field. If field already exist, update it with the  :visibility given, but the :type of the field cannot be changed after field is created. Doing so to protect legacy data. If such need persists, one shall explicit delete the field, and then create it with a different type.
      parameters:
        - name: name
          in: path
          description: name of the custom field to be created or updated
          required: true
          type: string
        - name: type
          in: path
          description: enum {'SHORT','LONG','INT','FLOAT'}
          required: true
          type: string
        - name: visibility
          in: path
          description: enum {'PRIVATE','PUBLIC'}
          required: true
          type: string
      tags:
        - Custom Fields
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /api/items/customFields/{name}/:
    delete:
      security:
        - JWT: []
      description: Remove a custom field on items. All values associated with this field will also be deleted.
      parameters:
        - name: name
          in: path
          description: name of the custom field to be deleted
          required: true
          type: string
      tags:
        - Custom Fields
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /api/items/{itemName}/{quantity}:
    post:
      security:
        - JWT: []
      description: Create a new item with {itemName} and {quantity}
      parameters:
        - name: itemName
          in: path
          description: item name
          required: true
          type: string
        - name: quantity
          in: path
          description: count of items
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            type: object
            # required: []
            properties:
              model:
                description: item model
                type: string
              description:
                description: describe the item in details
                type: string
              location:
                description: the location of the item
                type: string
              tags:
                description: an array of tags, each tag is a string
                type: object
              itemStatus:
                description: enum {"ACTIVE", "INACTIVE"}
                type: string
      tags:
        - Items
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /api/items/{itemId}:
    put:
      security:
        - JWT: []
      description: Update by overriding an existing item with {itemId}
      parameters:
        - name: itemId
          in: path
          description: count of items
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            type: object
            # required: []
            properties:
              name:
                description: item name
                type: string
              quantity:
                description: count of the item
                type: integer
              model:
                description: item model
                type: string
              description:
                description: describe the item in details
                type: string
              location:
                description: the location of the item
                type: string
              tags:
                description: an array of tags, each tag is a string
                type: object
              itemStatus:
                description: enum {"ACTIVE", "INACTIVE"}
                type: string
      tags:
        - Items
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /api/items/{itemId}/:
    delete:
      security:
        - JWT: []
      description: Remove an item
      parameters:
        - name: itemId
          in: path
          description: identifies an item
          required: true
          type: integer
      tags:
        - Items
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /api/tags/:
    delete:
      security:
        - JWT: []
      description: get all the tags
      tags:
        - Tags
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  # /api/tags/:
    post:
      security:
        - JWT: []
      description: create new tags from the array provided (will ignore those already existing).
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            # required: []
            properties:
              tags:
                description: array of tags, each as a string
                type: string
      tags:
        - Tags
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /api/tags/{tagId}/:
    delete:
      security:
        - JWT: []
      description: delete an existing tag with :tagId. Also removes
      parameters:
        - name: tagId
          in: path
          description: identifies an item
          required: true
          type: integer
      tags:
        - Tags
      responses:
        200:
          description: Response object
          schema:
            $ref: '#/definitions/ResponseWithObjectDataType'
        default:
          description: Error
          schema:
            $ref: '#/definitions/Error'
definitions:
...
  ResponseWithObjectDataType:
    type: object
    properties:
      status:
        description: "success"
        type: string
      data:
        description: An object
        type: object
  Error:
    type: object
    properties:
      status:
        description: "error"
        type: string
      error:
        description: Error object
        type: object